+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+**New**
+
+- Add alias between `cads_text_area` and `cads_textarea`
+
 **Fix**
 
 - Fix translation day, month, and year labels when using `cads_date_field`
@@ -7,6 +11,7 @@
 
 **Deprecations**
 
+- Log a deprecation warning when passing `additional_attributes` hash to `cads_text_field` or `cads_text_area`. These methods now support the Rails convention of passing the default options hash directly through to the element.
 - Log a deprecation warning when using the `type` argument for `Textarea` and `Select` components. Neither textarea nor select elements accept a type. This argument will be removed in a future version.
 
 ## v8.3.2